pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proof-of-process"
version = "0.1.0"
description = "Proof of Process — Enterprise-grade Project Analytics pipeline (ingest → charts → PDF)."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "PoP Team" }
]
keywords = ["project-management", "analytics", "matplotlib", "reportlab", "pipeline"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "matplotlib>=3.9",
  "reportlab>=4.2",
  "PyYAML>=6.0",
  "python-dateutil>=2.9",
  "scipy>=1.13",
  "unidecode>=1.3",
  "openpyxl>=3.1",
  "requests>=2.32",
]

[project.optional-dependencies]
gsheets = [
  "gspread>=6.0",
  "google-auth>=2.0",
]

[project.urls]
Homepage = "https://example.com/proof-of-process"

[project.scripts]
pop-run-pipeline = "proof_of_process.cli.run_pipeline:main"
pop-ingest-run = "proof_of_process.cli.ingest_then_run:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["proof_of_process*"]

